fix(clippy): satisfy pedantic linting
Clean the remaining clippy pedantic findings in clam-sigutil with mechanical changes: build-script iteration cleanup, accessor annotations, display-format simplifications, private-field naming, match-arm consolidation, and test assertion formatting.

Keep the crate's existing documentation-lint policy explicit at the crate level so command-line -D clippy::pedantic behaves consistently with Cargo.toml policy while still enforcing the behavioral and style lint fixes.

Validation: cargo clippy --all-targets --locked -- -D warnings -D clippy::pedantic; cargo clippy --no-default-features --lib --tests --locked -- -D warnings -D clippy::pedantic; cargo test --locked; cargo test --no-default-features --locked; cargo tree --no-default-features | rg openssl returned no matches

+++ src/signature/logical_sig/subsig/bytecmp/compset.rs
@@ -54,14 +54,17 @@ pub enum ComparisonSetParseError {
 }
 
 impl ComparisonSet {
+    #[must_use]
     pub fn symbol(&self) -> ComparisonOp {
         self.symbol
     }
 
+    #[must_use]
     pub fn value(&self) -> i64 {
         self.value
     }
 
+    #[must_use]
     pub fn encoding(&self) -> Encoding {
         self.encoding
     }
@@ -100,8 +103,7 @@ impl TryFrom<&[u8]> for ComparisonSet {
         } else {
             (
                 Encoding::Decimal,
-                parse_number_dec::<i64>(remainder).map_err(ComparisonSetParseError::ParseValue)?
-                    as i64,
+                parse_number_dec::<i64>(remainder).map_err(ComparisonSetParseError::ParseValue)?,
             )
         };
